Move post transform for third stage
