+++ sample-apps/radiology/lib/transforms/transforms.py
@@ -530,26 +530,3 @@ class NormalizeLabelsInDatasetd(MapTransform):
             d["label_names"] = new_label_names
             d[key].array = label
         return d
-
-
-class MergeAllPreds(MapTransform):
-    def __init__(self, keys: KeysCollection, allow_missing_keys: bool = False):
-        """
-        Merge all predictions to one channel
-
-        Args:
-            keys: The ``keys`` parameter will be used to get and set the actual data item to transform
-            label_names: all label names
-        """
-        super().__init__(keys, allow_missing_keys)
-
-    def __call__(self, data: Mapping[Hashable, NdarrayOrTensor]):
-        d: Dict = dict(data)
-        for idx, key in enumerate(self.key_iterator(d)):
-            if idx == 0:
-                merge_image = d[key]
-            else:
-                merge_image = merge_image + d[key]
-            # For labels that overlap keep the last label number only
-            merge_image[merge_image > d[key].max()] = d[key].max()
-        return merge_image
